Add files via upload
This is a Python code that generates a band name based on the user's inputs.

The first line prints out the message "Welcome to the Band Name Generator."
The next line prompts the user to input the name of the city they grew up in and stores the input in the variable "street". The "\n" character creates a new line after the prompt, making it more readable for the user.
The following line prompts the user to input their pet's name and stores the input in the variable "pet".
The last line prints out a message that combines the input values of the variables "street" and "pet" to create the band name, using the string concatenation operation (+). The final output is a string that says "Your band name could be (name of the city you grew up in) (name of your pet)".